web/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
python-multipart==0.0.22
python-jose[cryptography]==3.5.0
bcrypt==5.0.0
aiofiles==25.1.0
sqlalchemy==2.0.41
chromadb==1.5.4
requests==2.32.5
pypdf==6.8.0
python-docx==1.2.0
openpyxl==3.1.5
python-pptx
pytesseract==0.3.13
Pillow
faster-whisper==1.2.1
